Simplify finalizer results
